@@ -3,29 +3,31 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@chakra-ui/icons": "^1.0.12",
-    "@chakra-ui/react": "^1.3.4",
+    "@chakra-ui/react": "^3.0.0",
     "@emotion/react": "^11",
     "@emotion/styled": "^11",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "fontsource-roboto": "^4.0.0",
-    "framer-motion": "3.10.6",
-    "markdown-to-jsx": "^7.1.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-icons": "^4.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.2",
-    "react-syntax-highlighter": "^15.4.3",
-    "typescript": "^4.2.3",
-    "web-vitals": "^1.0.1"
+    "@fontsource/roboto": "^5.2.6",
+    "@testing-library/dom": "^10.4.0",
+    "@testing-library/jest-dom": "^6.6.0",
+    "@testing-library/react": "^16.1.0",
+    "@testing-library/user-event": "^14.5.0",
+    "framer-motion": "^11.0.0",
+    "lucide-react": "^0.525.0",
+    "markdown-to-jsx": "^7.5.0",
+    "next-themes": "^0.4.6",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
+    "react-icons": "^5.4.0",
+    "react-router-dom": "^6.30.0",
+    "react-scripts": "5.0.1",
+    "react-syntax-highlighter": "^15.6.0",
+    "typescript": "^5.8.0",
+    "web-vitals": "^5.0.0"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test",
+    "test": "react-scripts test --passWithNoTests",
     "eject": "react-scripts eject",
     "push-s3": "aws s3 sync build/ s3://grossi.life --acl public-read",
     "deploy": "yarn build && yarn push-s3"
@@ -49,10 +51,11 @@
     ]
   },
   "devDependencies": {
-    "@types/react-dom": "^17.0.2",
-    "@types/react-router-dom": "^5.1.7",
-    "@types/react-syntax-highlighter": "^13.5.0",
+    "@types/react": "^18.3.0",
+    "@types/react-dom": "^18.3.0",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/react-syntax-highlighter": "^15.5.13",
     "@types/styled-components": "^5.1.9",
-    "prettier": "^2.2.1"
+    "prettier": "^3.3.0"
   }
 }
